backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# Audio Processing
openai-whisper==20231117

# Embeddings & ML
sentence-transformers==2.2.2
transformers==4.35.2
torch
torchvision

# Vector Store - FAISS
faiss-cpu==1.7.4

# Database - PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Object Storage - MinIO
minio==7.2.0

# LLM
ollama

# Utilities
numpy==1.26.2
python-dotenv==1.0.0
aiofiles==23.2.1
